Network.AWS.ElasticSearch

Description

Amazon Elasticsearch Configuration Service

Use the Amazon Elasticsearch configuration API to create, configure, and manage Elasticsearch domains.

The endpoint for configuration service requests is region-specific: es.region .amazonaws.com. For example, es.us-east-1.amazonaws.com. For a current list of supported regions and endpoints, see Regions and Endpoints .

Service Configuration

elasticSearch :: Service

API version 2015-01-01 of the Amazon Elasticsearch Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ElasticSearch.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

An exception for missing / invalid input fields. Gives http status code of 400.

ResourceAlreadyExistsException

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

An exception for creating a resource that already exists. Gives http status code of 400.

BaseException

_BaseException :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred while processing the request.

DisabledOperationException

_DisabledOperationException :: AsError a => Getting (First ServiceError) a ServiceError

An error occured because the client wanted to access a not supported operation. Gives http status code of 409.

InternalException

_InternalException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception or failure (the failure is internal to the service) . Gives http status code of 500.

InvalidTypeException

_InvalidTypeException :: AsError a => Getting (First ServiceError) a ServiceError

An exception for trying to create or access sub-resource that is either invalid or not supported. Gives http status code of 409.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

An exception for accessing or deleting a resource that does not exist. Gives http status code of 400.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

An exception for trying to create more than allowed resources or sub-resources. Gives http status code of 409.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreateElasticsearchDomain

module Network.AWS.ElasticSearch.CreateElasticsearchDomain

RemoveTags

module Network.AWS.ElasticSearch.RemoveTags

DescribeElasticsearchDomains

module Network.AWS.ElasticSearch.DescribeElasticsearchDomains

ListElasticsearchInstanceTypes

module Network.AWS.ElasticSearch.ListElasticsearchInstanceTypes

DescribeElasticsearchDomain

module Network.AWS.ElasticSearch.DescribeElasticsearchDomain

ListDomainNames

module Network.AWS.ElasticSearch.ListDomainNames

DescribeElasticsearchInstanceTypeLimits

module Network.AWS.ElasticSearch.DescribeElasticsearchInstanceTypeLimits

DescribeElasticsearchDomainConfig

module Network.AWS.ElasticSearch.DescribeElasticsearchDomainConfig

DeleteElasticsearchDomain

module Network.AWS.ElasticSearch.DeleteElasticsearchDomain

UpdateElasticsearchDomainConfig

module Network.AWS.ElasticSearch.UpdateElasticsearchDomainConfig

ListElasticsearchVersions

module Network.AWS.ElasticSearch.ListElasticsearchVersions

AddTags

module Network.AWS.ElasticSearch.AddTags

ListTags

module Network.AWS.ElasticSearch.ListTags

Types

ESPartitionInstanceType

data ESPartitionInstanceType

Constructors

C4_2XLarge_Elasticsearch

C4_4XLarge_Elasticsearch

C4_8XLarge_Elasticsearch

C4_Large_Elasticsearch

C4_XLarge_Elasticsearch

D2_2XLarge_Elasticsearch

D2_4XLarge_Elasticsearch

D2_8XLarge_Elasticsearch

D2_XLarge_Elasticsearch

I2_2XLarge_Elasticsearch

I2_XLarge_Elasticsearch

M3_2XLarge_Elasticsearch

M3_Large_Elasticsearch

M3_Medium_Elasticsearch

M3_XLarge_Elasticsearch

M4_10XLarge_Elasticsearch

M4_2XLarge_Elasticsearch

M4_4XLarge_Elasticsearch

M4_Large_Elasticsearch

M4_XLarge_Elasticsearch

R3_2XLarge_Elasticsearch

R3_4XLarge_Elasticsearch

R3_8XLarge_Elasticsearch

R3_Large_Elasticsearch

R3_XLarge_Elasticsearch

R4_16XLarge_Elasticsearch

R4_2XLarge_Elasticsearch

R4_4XLarge_Elasticsearch

R4_8XLarge_Elasticsearch

R4_Large_Elasticsearch

R4_XLarge_Elasticsearch

T2_Medium_Elasticsearch

T2_Micro_Elasticsearch

T2_Small_Elasticsearch

OptionState

data OptionState

The state of a requested change. One of the following:

Processing: The request change is still in-process. * Active: The request change is processed and deployed to the Elasticsearch domain.

Constructors

Active

Processing

RequiresIndexDocuments

VolumeType

data VolumeType

The type of EBS volume, standard, gp2, or io1. See Configuring EBS-based Storage for more information.

Constructors

GP2

IO1

Standard

AccessPoliciesStatus

data AccessPoliciesStatus

The configured access rules for the domain's document and search endpoints, and the current status of those rules.

See: accessPoliciesStatus smart constructor.

accessPoliciesStatus

Arguments

:: Text
apsOptions

-> OptionStatus
apsStatus

-> AccessPoliciesStatus

Creates a value of AccessPoliciesStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apsOptions - The access policy configured for the Elasticsearch domain. Access policies may be resource-based, IP-based, or IAM-based. See Configuring Access Policies for more information.

apsStatus - The status of the access policy for the Elasticsearch domain. See OptionStatus for the status information that's included.

apsOptions :: Lens' AccessPoliciesStatus Text

The access policy configured for the Elasticsearch domain. Access policies may be resource-based, IP-based, or IAM-based. See Configuring Access Policies for more information.

apsStatus :: Lens' AccessPoliciesStatus OptionStatus

The status of the access policy for the Elasticsearch domain. See OptionStatus for the status information that's included.

AdditionalLimit

data AdditionalLimit

List of limits that are specific to a given InstanceType and for each of it's InstanceRole .

See: additionalLimit smart constructor.

additionalLimit :: AdditionalLimit

Creates a value of AdditionalLimit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

alLimitName - Name of Additional Limit is specific to a given InstanceType and for each of it's InstanceRole etc. Attributes and their details: * MaximumNumberOfDataNodesSupportedThis attribute will be present in Master node only to specify how much data nodes upto which given ESPartitionInstanceType can support as master node. * MaximumNumberOfDataNodesWithoutMasterNodeThis attribute will be present in Data node only to specify how much data nodes of given ESPartitionInstanceType upto which you don't need any master nodes to govern them.

alLimitValues - Value for given 'AdditionalLimit$LimitName' .

alLimitName :: Lens' AdditionalLimit (Maybe Text)

Name of Additional Limit is specific to a given InstanceType and for each of it's InstanceRole etc. Attributes and their details: * MaximumNumberOfDataNodesSupportedThis attribute will be present in Master node only to specify how much data nodes upto which given ESPartitionInstanceType can support as master node. * MaximumNumberOfDataNodesWithoutMasterNodeThis attribute will be present in Data node only to specify how much data nodes of given ESPartitionInstanceType upto which you don't need any master nodes to govern them.

alLimitValues :: Lens' AdditionalLimit [Text]

Value for given 'AdditionalLimit$LimitName' .

AdvancedOptionsStatus

data AdvancedOptionsStatus

Status of the advanced options for the specified Elasticsearch domain. Currently, the following advanced options are available:

Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true . See Configuration Advanced Options for more information. * Option to specify the percentage of heap space that is allocated to field data. By default, this setting is unbounded.

For more information, see Configuring Advanced Options .

See: advancedOptionsStatus smart constructor.

advancedOptionsStatus

Arguments

:: OptionStatus
aosStatus

-> AdvancedOptionsStatus

Creates a value of AdvancedOptionsStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aosOptions - Specifies the status of advanced options for the specified Elasticsearch domain.

aosStatus - Specifies the status of OptionStatus for advanced options for the specified Elasticsearch domain.

aosOptions :: Lens' AdvancedOptionsStatus (HashMap Text Text)

Specifies the status of advanced options for the specified Elasticsearch domain.

aosStatus :: Lens' AdvancedOptionsStatus OptionStatus

Specifies the status of OptionStatus for advanced options for the specified Elasticsearch domain.

DomainInfo

data DomainInfo

See: domainInfo smart constructor.

domainInfo :: DomainInfo

Creates a value of DomainInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diDomainName - Specifies the DomainName .

diDomainName :: Lens' DomainInfo (Maybe Text)

Specifies the DomainName .

EBSOptions

data EBSOptions

Options to enable, disable, and specify the properties of EBS storage volumes. For more information, see Configuring EBS-based Storage .

See: ebsOptions smart constructor.

ebsOptions :: EBSOptions

Creates a value of EBSOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eoVolumeSize - Integer to specify the size of an EBS volume.

eoIOPS - Specifies the IOPD for a Provisioned IOPS EBS volume (SSD).

eoVolumeType - Specifies the volume type for EBS-based storage.

eoEBSEnabled - Specifies whether EBS-based storage is enabled.

eoVolumeSize :: Lens' EBSOptions (Maybe Int)

Integer to specify the size of an EBS volume.

eoIOPS :: Lens' EBSOptions (Maybe Int)

Specifies the IOPD for a Provisioned IOPS EBS volume (SSD).

eoVolumeType :: Lens' EBSOptions (Maybe VolumeType)

Specifies the volume type for EBS-based storage.

eoEBSEnabled :: Lens' EBSOptions (Maybe Bool)

Specifies whether EBS-based storage is enabled.

EBSOptionsStatus

data EBSOptionsStatus

Status of the EBS options for the specified Elasticsearch domain.

See: ebsOptionsStatus smart constructor.

ebsOptionsStatus

Arguments

:: EBSOptions
eosOptions

-> OptionStatus
eosStatus

-> EBSOptionsStatus

Creates a value of EBSOptionsStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eosOptions - Specifies the EBS options for the specified Elasticsearch domain.

eosStatus - Specifies the status of the EBS options for the specified Elasticsearch domain.

eosOptions :: Lens' EBSOptionsStatus EBSOptions

Specifies the EBS options for the specified Elasticsearch domain.

eosStatus :: Lens' EBSOptionsStatus OptionStatus

Specifies the status of the EBS options for the specified Elasticsearch domain.

ElasticsearchClusterConfig

data ElasticsearchClusterConfig

Specifies the configuration for the domain cluster, such as the type and number of instances.

See: elasticsearchClusterConfig smart constructor.

elasticsearchClusterConfig :: ElasticsearchClusterConfig

Creates a value of ElasticsearchClusterConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eccDedicatedMasterCount - Total number of dedicated master nodes, active and on standby, for the cluster.

eccDedicatedMasterType - The instance type for a dedicated master node.

eccDedicatedMasterEnabled - A boolean value to indicate whether a dedicated master node is enabled. See About Dedicated Master Nodes for more information.

eccInstanceCount - The number of instances in the specified domain cluster.

eccZoneAwarenessEnabled - A boolean value to indicate whether zone awareness is enabled. See About Zone Awareness for more information.

eccInstanceType - The instance type for an Elasticsearch cluster.

eccDedicatedMasterCount :: Lens' ElasticsearchClusterConfig (Maybe Int)

Total number of dedicated master nodes, active and on standby, for the cluster.

eccDedicatedMasterType :: Lens' ElasticsearchClusterConfig (Maybe ESPartitionInstanceType)

The instance type for a dedicated master node.

eccDedicatedMasterEnabled :: Lens' ElasticsearchClusterConfig (Maybe Bool)

A boolean value to indicate whether a dedicated master node is enabled. See About Dedicated Master Nodes for more information.

eccInstanceCount :: Lens' ElasticsearchClusterConfig (Maybe Int)

The number of instances in the specified domain cluster.

eccZoneAwarenessEnabled :: Lens' ElasticsearchClusterConfig (Maybe Bool)

A boolean value to indicate whether zone awareness is enabled. See About Zone Awareness for more information.

eccInstanceType :: Lens' ElasticsearchClusterConfig (Maybe ESPartitionInstanceType)

The instance type for an Elasticsearch cluster.

ElasticsearchClusterConfigStatus

data ElasticsearchClusterConfigStatus

Specifies the configuration status for the specified Elasticsearch domain.

See: elasticsearchClusterConfigStatus smart constructor.

elasticsearchClusterConfigStatus

Arguments

:: ElasticsearchClusterConfig
eccsOptions

-> OptionStatus
eccsStatus

-> ElasticsearchClusterConfigStatus

Creates a value of ElasticsearchClusterConfigStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eccsOptions - Specifies the cluster configuration for the specified Elasticsearch domain.

eccsStatus - Specifies the status of the configuration for the specified Elasticsearch domain.

eccsOptions :: Lens' ElasticsearchClusterConfigStatus ElasticsearchClusterConfig

Specifies the cluster configuration for the specified Elasticsearch domain.

eccsStatus :: Lens' ElasticsearchClusterConfigStatus OptionStatus

Specifies the status of the configuration for the specified Elasticsearch domain.

ElasticsearchDomainConfig

data ElasticsearchDomainConfig

The configuration of an Elasticsearch domain.

See: elasticsearchDomainConfig smart constructor.

elasticsearchDomainConfig :: ElasticsearchDomainConfig

Creates a value of ElasticsearchDomainConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

edcEBSOptions - Specifies the EBSOptions for the Elasticsearch domain.

edcAccessPolicies - IAM access policy as a JSON-formatted string.

edcElasticsearchClusterConfig - Specifies the ElasticsearchClusterConfig for the Elasticsearch domain.

edcSnapshotOptions - Specifies the SnapshotOptions for the Elasticsearch domain.

edcAdvancedOptions - Specifies the AdvancedOptions for the domain. See Configuring Advanced Options for more information.

edcElasticsearchVersion - String of format X.Y to specify version for the Elasticsearch domain.

edcEBSOptions :: Lens' ElasticsearchDomainConfig (Maybe EBSOptionsStatus)

Specifies the EBSOptions for the Elasticsearch domain.

edcAccessPolicies :: Lens' ElasticsearchDomainConfig (Maybe AccessPoliciesStatus)

IAM access policy as a JSON-formatted string.

edcElasticsearchClusterConfig :: Lens' ElasticsearchDomainConfig (Maybe ElasticsearchClusterConfigStatus)

Specifies the ElasticsearchClusterConfig for the Elasticsearch domain.

edcSnapshotOptions :: Lens' ElasticsearchDomainConfig (Maybe SnapshotOptionsStatus)

Specifies the SnapshotOptions for the Elasticsearch domain.

edcAdvancedOptions :: Lens' ElasticsearchDomainConfig (Maybe AdvancedOptionsStatus)

Specifies the AdvancedOptions for the domain. See Configuring Advanced Options for more information.

edcElasticsearchVersion :: Lens' ElasticsearchDomainConfig (Maybe ElasticsearchVersionStatus)

String of format X.Y to specify version for the Elasticsearch domain.

ElasticsearchDomainStatus

data ElasticsearchDomainStatus

The current status of an Elasticsearch domain.

See: elasticsearchDomainStatus smart constructor.

elasticsearchDomainStatus

Arguments

:: Text
edsDomainId

-> Text
edsDomainName

-> Text
edsARN

-> ElasticsearchClusterConfig
edsElasticsearchClusterConfig

-> ElasticsearchDomainStatus

Creates a value of ElasticsearchDomainStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

edsEBSOptions - The EBSOptions for the specified domain. See Configuring EBS-based Storage for more information.

edsAccessPolicies - IAM access policy as a JSON-formatted string.

edsCreated - The domain creation status. True if the creation of an Elasticsearch domain is complete. False if domain creation is still in progress.

edsSnapshotOptions - Specifies the status of the SnapshotOptions

edsDeleted - The domain deletion status. True if a delete request has been received for the domain but resource cleanup is still in progress. False if the domain has not been deleted. Once domain deletion is complete, the status of the domain is no longer returned.

edsProcessing - The status of the Elasticsearch domain configuration. True if Amazon Elasticsearch Service is processing configuration changes. False if the configuration is active.

edsEndpoint - The Elasticsearch domain endpoint that you use to submit index and search requests.

edsAdvancedOptions - Specifies the status of the AdvancedOptions

edsElasticsearchVersion - Undocumented member.

edsDomainId - The unique identifier for the specified Elasticsearch domain.

edsDomainName - The name of an Elasticsearch domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

edsARN - The Amazon resource name (ARN) of an Elasticsearch domain. See Identifiers for IAM Entities in Using AWS Identity and Access Management for more information.

edsElasticsearchClusterConfig - The type and number of instances in the domain cluster.

edsEBSOptions :: Lens' ElasticsearchDomainStatus (Maybe EBSOptions)

The EBSOptions for the specified domain. See Configuring EBS-based Storage for more information.

edsAccessPolicies :: Lens' ElasticsearchDomainStatus (Maybe Text)

IAM access policy as a JSON-formatted string.

edsCreated :: Lens' ElasticsearchDomainStatus (Maybe Bool)

The domain creation status. True if the creation of an Elasticsearch domain is complete. False if domain creation is still in progress.

edsSnapshotOptions :: Lens' ElasticsearchDomainStatus (Maybe SnapshotOptions)

Specifies the status of the SnapshotOptions

edsDeleted :: Lens' ElasticsearchDomainStatus (Maybe Bool)

The domain deletion status. True if a delete request has been received for the domain but resource cleanup is still in progress. False if the domain has not been deleted. Once domain deletion is complete, the status of the domain is no longer returned.

edsProcessing :: Lens' ElasticsearchDomainStatus (Maybe Bool)

The status of the Elasticsearch domain configuration. True if Amazon Elasticsearch Service is processing configuration changes. False if the configuration is active.

edsEndpoint :: Lens' ElasticsearchDomainStatus (Maybe Text)

The Elasticsearch domain endpoint that you use to submit index and search requests.

edsAdvancedOptions :: Lens' ElasticsearchDomainStatus (HashMap Text Text)

Specifies the status of the AdvancedOptions

edsElasticsearchVersion :: Lens' ElasticsearchDomainStatus (Maybe Text)

Undocumented member.

edsDomainId :: Lens' ElasticsearchDomainStatus Text

The unique identifier for the specified Elasticsearch domain.

edsDomainName :: Lens' ElasticsearchDomainStatus Text

The name of an Elasticsearch domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

edsARN :: Lens' ElasticsearchDomainStatus Text

The Amazon resource name (ARN) of an Elasticsearch domain. See Identifiers for IAM Entities in Using AWS Identity and Access Management for more information.

edsElasticsearchClusterConfig :: Lens' ElasticsearchDomainStatus ElasticsearchClusterConfig

The type and number of instances in the domain cluster.

ElasticsearchVersionStatus

data ElasticsearchVersionStatus

Status of the Elasticsearch version options for the specified Elasticsearch domain.

See: elasticsearchVersionStatus smart constructor.

elasticsearchVersionStatus

Arguments

:: Text
evsOptions

-> OptionStatus
evsStatus

-> ElasticsearchVersionStatus

Creates a value of ElasticsearchVersionStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

evsOptions - Specifies the Elasticsearch version for the specified Elasticsearch domain.

evsStatus - Specifies the status of the Elasticsearch version options for the specified Elasticsearch domain.

evsOptions :: Lens' ElasticsearchVersionStatus Text

Specifies the Elasticsearch version for the specified Elasticsearch domain.

evsStatus :: Lens' ElasticsearchVersionStatus OptionStatus

Specifies the status of the Elasticsearch version options for the specified Elasticsearch domain.

InstanceCountLimits

data InstanceCountLimits

InstanceCountLimits represents the limits on number of instances that be created in Amazon Elasticsearch for given InstanceType.

See: instanceCountLimits smart constructor.

instanceCountLimits :: InstanceCountLimits

Creates a value of InstanceCountLimits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iclMaximumInstanceCount - Undocumented member.

iclMinimumInstanceCount - Undocumented member.

iclMaximumInstanceCount :: Lens' InstanceCountLimits (Maybe Int)

Undocumented member.

iclMinimumInstanceCount :: Lens' InstanceCountLimits (Maybe Int)

Undocumented member.

InstanceLimits

data InstanceLimits

InstanceLimits represents the list of instance related attributes that are available for given InstanceType.

See: instanceLimits smart constructor.

instanceLimits :: InstanceLimits

Creates a value of InstanceLimits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ilInstanceCountLimits - Undocumented member.

ilInstanceCountLimits :: Lens' InstanceLimits (Maybe InstanceCountLimits)

Undocumented member.

Limits

data Limits

Limits for given InstanceType and for each of it's role.

Limits contains following 'StorageTypes,' InstanceLimits and AdditionalLimits

See: limits smart constructor.

limits :: Limits

Creates a value of Limits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lInstanceLimits - Undocumented member.

lAdditionalLimits - List of additional limits that are specific to a given InstanceType and for each of it's InstanceRole .

lStorageTypes - StorageType represents the list of storage related types and attributes that are available for given InstanceType.

lInstanceLimits :: Lens' Limits (Maybe InstanceLimits)

Undocumented member.

lAdditionalLimits :: Lens' Limits [AdditionalLimit]

List of additional limits that are specific to a given InstanceType and for each of it's InstanceRole .

lStorageTypes :: Lens' Limits [StorageType]

StorageType represents the list of storage related types and attributes that are available for given InstanceType.

OptionStatus

data OptionStatus

Provides the current status of the entity.

See: optionStatus smart constructor.

optionStatus

Arguments

:: UTCTime
osCreationDate

-> UTCTime
osUpdateDate

-> OptionState
osState

-> OptionStatus

Creates a value of OptionStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

osPendingDeletion - Indicates whether the Elasticsearch domain is being deleted.

osUpdateVersion - Specifies the latest version for the entity.

osCreationDate - Timestamp which tells the creation date for the entity.

osUpdateDate - Timestamp which tells the last updated time for the entity.

osState - Provides the OptionState for the Elasticsearch domain.

osPendingDeletion :: Lens' OptionStatus (Maybe Bool)

Indicates whether the Elasticsearch domain is being deleted.

osUpdateVersion :: Lens' OptionStatus (Maybe Natural)

Specifies the latest version for the entity.

osCreationDate :: Lens' OptionStatus UTCTime

Timestamp which tells the creation date for the entity.

osUpdateDate :: Lens' OptionStatus UTCTime

Timestamp which tells the last updated time for the entity.

osState :: Lens' OptionStatus OptionState

Provides the OptionState for the Elasticsearch domain.

SnapshotOptions

data SnapshotOptions

Specifies the time, in UTC format, when the service takes a daily automated snapshot of the specified Elasticsearch domain. Default value is 0 hours.

See: snapshotOptions smart constructor.

snapshotOptions :: SnapshotOptions

Creates a value of SnapshotOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

soAutomatedSnapshotStartHour - Specifies the time, in UTC format, when the service takes a daily automated snapshot of the specified Elasticsearch domain. Default value is 0 hours.

soAutomatedSnapshotStartHour :: Lens' SnapshotOptions (Maybe Int)

Specifies the time, in UTC format, when the service takes a daily automated snapshot of the specified Elasticsearch domain. Default value is 0 hours.

SnapshotOptionsStatus

data SnapshotOptionsStatus

Status of a daily automated snapshot.

See: snapshotOptionsStatus smart constructor.

snapshotOptionsStatus

Arguments

:: SnapshotOptions
sosOptions

-> OptionStatus
sosStatus

-> SnapshotOptionsStatus

Creates a value of SnapshotOptionsStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sosOptions - Specifies the daily snapshot options specified for the Elasticsearch domain.

sosStatus - Specifies the status of a daily automated snapshot.

sosOptions :: Lens' SnapshotOptionsStatus SnapshotOptions

Specifies the daily snapshot options specified for the Elasticsearch domain.

sosStatus :: Lens' SnapshotOptionsStatus OptionStatus

Specifies the status of a daily automated snapshot.

StorageType

data StorageType

StorageTypes represents the list of storage related types and their attributes that are available for given InstanceType.

See: storageType smart constructor.

storageType :: StorageType

Creates a value of StorageType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stStorageTypeLimits - List of limits that are applicable for given storage type.

stStorageSubTypeName - Undocumented member.

stStorageTypeName - Undocumented member.

stStorageTypeLimits :: Lens' StorageType [StorageTypeLimit]

List of limits that are applicable for given storage type.

stStorageSubTypeName :: Lens' StorageType (Maybe Text)

Undocumented member.

stStorageTypeName :: Lens' StorageType (Maybe Text)

Undocumented member.

StorageTypeLimit

data StorageTypeLimit

Limits that are applicable for given storage type.

See: storageTypeLimit smart constructor.

storageTypeLimit :: StorageTypeLimit

Creates a value of StorageTypeLimit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stlLimitName - Name of storage limits that are applicable for given storage type. If StorageType is ebs, following storage options are applicable * MinimumVolumeSizeMinimum amount of volume size that is applicable for given storage type.It can be empty if it is not applicable. * MaximumVolumeSizeMaximum amount of volume size that is applicable for given storage type.It can be empty if it is not applicable. * MaximumIopsMaximum amount of Iops that is applicable for given storage type.It can be empty if it is not applicable. * MinimumIopsMinimum amount of Iops that is applicable for given storage type.It can be empty if it is not applicable.

stlLimitValues - Values for the 'StorageTypeLimit$LimitName' .

stlLimitName :: Lens' StorageTypeLimit (Maybe Text)

Name of storage limits that are applicable for given storage type. If StorageType is ebs, following storage options are applicable * MinimumVolumeSizeMinimum amount of volume size that is applicable for given storage type.It can be empty if it is not applicable. * MaximumVolumeSizeMaximum amount of volume size that is applicable for given storage type.It can be empty if it is not applicable. * MaximumIopsMaximum amount of Iops that is applicable for given storage type.It can be empty if it is not applicable. * MinimumIopsMinimum amount of Iops that is applicable for given storage type.It can be empty if it is not applicable.

stlLimitValues :: Lens' StorageTypeLimit [Text]

Values for the 'StorageTypeLimit$LimitName' .

Tag

data Tag

Specifies a key value pair for a resource tag.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - Specifies the TagKey , the name of the tag. Tag keys must be unique for the Elasticsearch domain to which they are attached.

tagValue - Specifies the TagValue , the value assigned to the corresponding tag key. Tag values can be null and do not have to be unique in a tag set. For example, you can have a key value pair in a tag set of project : Trinity and cost-center : Trinity

tagKey :: Lens' Tag Text

Specifies the TagKey , the name of the tag. Tag keys must be unique for the Elasticsearch domain to which they are attached.

tagValue :: Lens' Tag Text

Specifies the TagValue , the value assigned to the corresponding tag key. Tag values can be null and do not have to be unique in a tag set. For example, you can have a key value pair in a tag set of project : Trinity and cost-center : Trinity