Network.AWS.ElasticSearch.DeleteElasticsearchDomain

Description

Permanently deletes the specified Elasticsearch domain and all of its data. Once a domain is deleted, it cannot be recovered.

Creating a Request

deleteElasticsearchDomain

Arguments

:: Text
dDomainName

-> DeleteElasticsearchDomain

Creates a value of DeleteElasticsearchDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dDomainName - The name of the Elasticsearch domain that you want to permanently delete.

data DeleteElasticsearchDomain

Container for the parameters to the DeleteElasticsearchDomain operation. Specifies the name of the Elasticsearch domain that you want to delete.

See: deleteElasticsearchDomain smart constructor.

Request Lenses

dDomainName :: Lens' DeleteElasticsearchDomain Text

The name of the Elasticsearch domain that you want to permanently delete.

Destructuring the Response

deleteElasticsearchDomainResponse

Arguments

:: Int
delrsResponseStatus

-> DeleteElasticsearchDomainResponse

Creates a value of DeleteElasticsearchDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delrsDomainStatus - The status of the Elasticsearch domain being deleted.

delrsResponseStatus - -- | The response status code.

data DeleteElasticsearchDomainResponse

The result of a DeleteElasticsearchDomain request. Contains the status of the pending deletion, or no status if the domain and all of its resources have been deleted.

See: deleteElasticsearchDomainResponse smart constructor.

Response Lenses

delrsDomainStatus :: Lens' DeleteElasticsearchDomainResponse (Maybe ElasticsearchDomainStatus)

The status of the Elasticsearch domain being deleted.

delrsResponseStatus :: Lens' DeleteElasticsearchDomainResponse Int

- | The response status code.