Network.AWS.ElasticSearch.DescribeElasticsearchDomain

Description

Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.

Creating a Request

describeElasticsearchDomain

Arguments

:: Text
dedDomainName

-> DescribeElasticsearchDomain

Creates a value of DescribeElasticsearchDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedDomainName - The name of the Elasticsearch domain for which you want information.

data DescribeElasticsearchDomain

Container for the parameters to the DescribeElasticsearchDomain operation.

See: describeElasticsearchDomain smart constructor.

Request Lenses

dedDomainName :: Lens' DescribeElasticsearchDomain Text

The name of the Elasticsearch domain for which you want information.

Destructuring the Response

describeElasticsearchDomainResponse

Arguments

:: Int
dedrsResponseStatus

-> ElasticsearchDomainStatus
dedrsDomainStatus

-> DescribeElasticsearchDomainResponse

Creates a value of DescribeElasticsearchDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedrsResponseStatus - -- | The response status code.

dedrsDomainStatus - The current status of the Elasticsearch domain.

data DescribeElasticsearchDomainResponse

The result of a DescribeElasticsearchDomain request. Contains the status of the domain specified in the request.

See: describeElasticsearchDomainResponse smart constructor.

Response Lenses

dedrsResponseStatus :: Lens' DescribeElasticsearchDomainResponse Int

- | The response status code.

dedrsDomainStatus :: Lens' DescribeElasticsearchDomainResponse ElasticsearchDomainStatus

The current status of the Elasticsearch domain.