Network.AWS.ElasticSearch.DescribeElasticsearchDomainConfig

Description

Provides cluster configuration information about the specified Elasticsearch domain, such as the state, creation date, update version, and update date for cluster options.

Creating a Request

describeElasticsearchDomainConfig

Arguments

:: Text
dedcDomainName

-> DescribeElasticsearchDomainConfig

Creates a value of DescribeElasticsearchDomainConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedcDomainName - The Elasticsearch domain that you want to get information about.

data DescribeElasticsearchDomainConfig

Container for the parameters to the DescribeElasticsearchDomainConfig operation. Specifies the domain name for which you want configuration information.

See: describeElasticsearchDomainConfig smart constructor.

Request Lenses

dedcDomainName :: Lens' DescribeElasticsearchDomainConfig Text

The Elasticsearch domain that you want to get information about.

Destructuring the Response

describeElasticsearchDomainConfigResponse

Arguments

:: Int
dedcrsResponseStatus

-> ElasticsearchDomainConfig
dedcrsDomainConfig

-> DescribeElasticsearchDomainConfigResponse

Creates a value of DescribeElasticsearchDomainConfigResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedcrsResponseStatus - -- | The response status code.

dedcrsDomainConfig - The configuration information of the domain requested in the DescribeElasticsearchDomainConfig request.

data DescribeElasticsearchDomainConfigResponse

The result of a DescribeElasticsearchDomainConfig request. Contains the configuration information of the requested domain.

See: describeElasticsearchDomainConfigResponse smart constructor.

Response Lenses

dedcrsResponseStatus :: Lens' DescribeElasticsearchDomainConfigResponse Int

- | The response status code.

dedcrsDomainConfig :: Lens' DescribeElasticsearchDomainConfigResponse ElasticsearchDomainConfig

The configuration information of the domain requested in the DescribeElasticsearchDomainConfig request.