Network.AWS.ElasticSearch.DescribeElasticsearchDomains

Description

Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.

Creating a Request

describeElasticsearchDomains :: DescribeElasticsearchDomains

Creates a value of DescribeElasticsearchDomains with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedDomainNames - The Elasticsearch domains for which you want information.

data DescribeElasticsearchDomains

Container for the parameters to the DescribeElasticsearchDomains operation. By default, the API returns the status of all Elasticsearch domains.

See: describeElasticsearchDomains smart constructor.

Request Lenses

dedDomainNames :: Lens' DescribeElasticsearchDomains [Text]

The Elasticsearch domains for which you want information.

Destructuring the Response

describeElasticsearchDomainsResponse

Arguments

:: Int
drsResponseStatus

-> DescribeElasticsearchDomainsResponse

Creates a value of DescribeElasticsearchDomainsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsResponseStatus - -- | The response status code.

drsDomainStatusList - The status of the domains requested in the DescribeElasticsearchDomains request.

data DescribeElasticsearchDomainsResponse

The result of a DescribeElasticsearchDomains request. Contains the status of the specified domains or all domains owned by the account.

See: describeElasticsearchDomainsResponse smart constructor.

Response Lenses

drsResponseStatus :: Lens' DescribeElasticsearchDomainsResponse Int

- | The response status code.

drsDomainStatusList :: Lens' DescribeElasticsearchDomainsResponse [ElasticsearchDomainStatus]

The status of the domains requested in the DescribeElasticsearchDomains request.