Network.AWS.ElasticSearch.DescribeElasticsearchInstanceTypeLimits

Description

Describe Elasticsearch Limits for a given InstanceType and ElasticsearchVersion. When modifying existing Domain, specify the DomainName to know what Limits are supported for modifying.

Creating a Request

describeElasticsearchInstanceTypeLimits

Arguments

:: ESPartitionInstanceType
deitlInstanceType

-> Text
deitlElasticsearchVersion

-> DescribeElasticsearchInstanceTypeLimits

Creates a value of DescribeElasticsearchInstanceTypeLimits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deitlDomainName - DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for Elasticsearch Limits for existing domain.

deitlInstanceType - The instance type for an Elasticsearch cluster for which Elasticsearch Limits are needed.

deitlElasticsearchVersion - Version of Elasticsearch for which Limits are needed.

data DescribeElasticsearchInstanceTypeLimits

Container for the parameters to DescribeElasticsearchInstanceTypeLimits operation.

See: describeElasticsearchInstanceTypeLimits smart constructor.

Request Lenses

deitlDomainName :: Lens' DescribeElasticsearchInstanceTypeLimits (Maybe Text)

DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for Elasticsearch Limits for existing domain.

deitlInstanceType :: Lens' DescribeElasticsearchInstanceTypeLimits ESPartitionInstanceType

The instance type for an Elasticsearch cluster for which Elasticsearch Limits are needed.

deitlElasticsearchVersion :: Lens' DescribeElasticsearchInstanceTypeLimits Text

Version of Elasticsearch for which Limits are needed.

Destructuring the Response

describeElasticsearchInstanceTypeLimitsResponse

Arguments

:: Int
deitlrsResponseStatus

-> DescribeElasticsearchInstanceTypeLimitsResponse

Creates a value of DescribeElasticsearchInstanceTypeLimitsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deitlrsLimitsByRole - Undocumented member.

deitlrsResponseStatus - -- | The response status code.

data DescribeElasticsearchInstanceTypeLimitsResponse

Container for the parameters received from DescribeElasticsearchInstanceTypeLimits operation.

See: describeElasticsearchInstanceTypeLimitsResponse smart constructor.

Response Lenses

deitlrsLimitsByRole :: Lens' DescribeElasticsearchInstanceTypeLimitsResponse (HashMap Text Limits)

Undocumented member.

deitlrsResponseStatus :: Lens' DescribeElasticsearchInstanceTypeLimitsResponse Int

- | The response status code.