Network.AWS.ElasticSearch.ListElasticsearchInstanceTypes

Description

List all Elasticsearch instance types that are supported for given ElasticsearchVersion

Creating a Request

listElasticsearchInstanceTypes

Arguments

:: Text
leitElasticsearchVersion

-> ListElasticsearchInstanceTypes

Creates a value of ListElasticsearchInstanceTypes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

leitNextToken - NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

leitDomainName - DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for list of available Elasticsearch instance types when modifying existing domain.

leitMaxResults - Set this value to limit the number of results returned. Value provided must be greater than 30 else it wont be honored.

leitElasticsearchVersion - Version of Elasticsearch for which list of supported elasticsearch instance types are needed.

data ListElasticsearchInstanceTypes

Container for the parameters to the ListElasticsearchInstanceTypes operation.

See: listElasticsearchInstanceTypes smart constructor.

Request Lenses

leitNextToken :: Lens' ListElasticsearchInstanceTypes (Maybe Text)

NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

leitDomainName :: Lens' ListElasticsearchInstanceTypes (Maybe Text)

DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for list of available Elasticsearch instance types when modifying existing domain.

leitMaxResults :: Lens' ListElasticsearchInstanceTypes (Maybe Int)

Set this value to limit the number of results returned. Value provided must be greater than 30 else it wont be honored.

leitElasticsearchVersion :: Lens' ListElasticsearchInstanceTypes Text

Version of Elasticsearch for which list of supported elasticsearch instance types are needed.

Destructuring the Response

listElasticsearchInstanceTypesResponse

Arguments

:: Int
leitrsResponseStatus

-> ListElasticsearchInstanceTypesResponse

Creates a value of ListElasticsearchInstanceTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

leitrsElasticsearchInstanceTypes - List of instance types supported by Amazon Elasticsearch service for given ElasticsearchVersion

leitrsNextToken - In case if there are more results available NextToken would be present, make further request to the same API with received NextToken to paginate remaining results.

leitrsResponseStatus - -- | The response status code.

data ListElasticsearchInstanceTypesResponse

Container for the parameters returned by ListElasticsearchInstanceTypes operation.

See: listElasticsearchInstanceTypesResponse smart constructor.

Response Lenses

leitrsElasticsearchInstanceTypes :: Lens' ListElasticsearchInstanceTypesResponse [ESPartitionInstanceType]

List of instance types supported by Amazon Elasticsearch service for given ElasticsearchVersion

leitrsNextToken :: Lens' ListElasticsearchInstanceTypesResponse (Maybe Text)

In case if there are more results available NextToken would be present, make further request to the same API with received NextToken to paginate remaining results.

leitrsResponseStatus :: Lens' ListElasticsearchInstanceTypesResponse Int

- | The response status code.