Network.AWS.ElasticSearch.ListElasticsearchVersions

Description

List all supported Elasticsearch versions

Creating a Request

listElasticsearchVersions :: ListElasticsearchVersions

Creates a value of ListElasticsearchVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

levNextToken - Undocumented member.

levMaxResults - Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.

data ListElasticsearchVersions

Container for the parameters to the ListElasticsearchVersions operation. Use MaxResults to control the maximum number of results to retrieve in a single call.

Use NextToken in response to retrieve more results. If the received response does not contain a NextToken, then there are no more results to retrieve.

See: listElasticsearchVersions smart constructor.

Request Lenses

levNextToken :: Lens' ListElasticsearchVersions (Maybe Text)

Undocumented member.

levMaxResults :: Lens' ListElasticsearchVersions (Maybe Int)

Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.

Destructuring the Response

listElasticsearchVersionsResponse

Arguments

:: Int
levrsResponseStatus

-> ListElasticsearchVersionsResponse

Creates a value of ListElasticsearchVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

levrsNextToken - Undocumented member.

levrsElasticsearchVersions - Undocumented member.

levrsResponseStatus - -- | The response status code.

data ListElasticsearchVersionsResponse

Container for the parameters for response received from ListElasticsearchVersions operation.

See: listElasticsearchVersionsResponse smart constructor.

Response Lenses

levrsNextToken :: Lens' ListElasticsearchVersionsResponse (Maybe Text)

Undocumented member.

levrsElasticsearchVersions :: Lens' ListElasticsearchVersionsResponse [Text]

Undocumented member.

levrsResponseStatus :: Lens' ListElasticsearchVersionsResponse Int

- | The response status code.