Network.AWS.ElasticSearch.UpdateElasticsearchDomainConfig

Description

Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances.

Creating a Request

updateElasticsearchDomainConfig

Arguments

:: Text
uedcDomainName

-> UpdateElasticsearchDomainConfig

Creates a value of UpdateElasticsearchDomainConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uedcEBSOptions - Specify the type and size of the EBS volume that you want to use.

uedcAccessPolicies - IAM access policy as a JSON-formatted string.

uedcElasticsearchClusterConfig - The type and number of instances to instantiate for the domain cluster.

uedcSnapshotOptions - Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

uedcAdvancedOptions - Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true . See Configuration Advanced Options for more information.

uedcDomainName - The name of the Elasticsearch domain that you are updating.

data UpdateElasticsearchDomainConfig

Container for the parameters to the UpdateElasticsearchDomain operation. Specifies the type and number of instances in the domain cluster.

See: updateElasticsearchDomainConfig smart constructor.

Request Lenses

uedcEBSOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe EBSOptions)

Specify the type and size of the EBS volume that you want to use.

uedcAccessPolicies :: Lens' UpdateElasticsearchDomainConfig (Maybe Text)

IAM access policy as a JSON-formatted string.

uedcElasticsearchClusterConfig :: Lens' UpdateElasticsearchDomainConfig (Maybe ElasticsearchClusterConfig)

The type and number of instances to instantiate for the domain cluster.

uedcSnapshotOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe SnapshotOptions)

Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

uedcAdvancedOptions :: Lens' UpdateElasticsearchDomainConfig (HashMap Text Text)

Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true . See Configuration Advanced Options for more information.

uedcDomainName :: Lens' UpdateElasticsearchDomainConfig Text

The name of the Elasticsearch domain that you are updating.

Destructuring the Response

updateElasticsearchDomainConfigResponse

Arguments

:: Int
uedcrsResponseStatus

-> ElasticsearchDomainConfig
uedcrsDomainConfig

-> UpdateElasticsearchDomainConfigResponse

Creates a value of UpdateElasticsearchDomainConfigResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uedcrsResponseStatus - -- | The response status code.

uedcrsDomainConfig - The status of the updated Elasticsearch domain.

data UpdateElasticsearchDomainConfigResponse

The result of an UpdateElasticsearchDomain request. Contains the status of the Elasticsearch domain being updated.

See: updateElasticsearchDomainConfigResponse smart constructor.

Response Lenses

uedcrsResponseStatus :: Lens' UpdateElasticsearchDomainConfigResponse Int

- | The response status code.

uedcrsDomainConfig :: Lens' UpdateElasticsearchDomainConfigResponse ElasticsearchDomainConfig

The status of the updated Elasticsearch domain.